+++ src/time_split/_frontend/_to_string.py
@@ -1,3 +1,4 @@
+import datetime
 from typing import Any, Literal, overload
 
 from pandas import Timestamp
@@ -10,9 +11,9 @@ from ..types import DatetimeSplitBounds, DatetimeTypes, ExpandLimits
 
 @overload
 def to_string(
-    bounds: DatetimeTypes,
-    __ignored0: DatetimeTypes,
-    __ignored1: DatetimeTypes,
+    bounds: DatetimeSplitBounds | tuple[DatetimeTypes, DatetimeTypes, DatetimeTypes],
+    __ignored0: Literal[None] = None,
+    __ignored1: Literal[None] = None,
     /,
     *,
     format: str | None = None,
@@ -21,9 +22,9 @@ def to_string(
 
 @overload
 def to_string(
-    start: tuple[DatetimeTypes, DatetimeTypes, DatetimeTypes],
-    mid: Literal[None] = None,
-    end: Literal[None] = None,
+    start: DatetimeTypes,
+    mid: DatetimeTypes,
+    end: DatetimeTypes,
     /,
     *,
     format: str | None = None,
@@ -31,7 +32,7 @@ def to_string(
 
 
 def to_string(
-    bounds: DatetimeTypes | DatetimeSplitBounds | tuple[DatetimeTypes, DatetimeTypes, DatetimeTypes],
+    start_or_bounds: DatetimeTypes | DatetimeSplitBounds | tuple[DatetimeTypes, DatetimeTypes, DatetimeTypes],
     mid: DatetimeTypes | None = None,
     end: DatetimeTypes | None = None,
     /,
@@ -40,15 +41,10 @@ def to_string(
 ) -> str:
     """Pretty-print a fold.
 
-    .. code-block:: python
-       :caption: Sample output.
-
-       ('2021-12-30' <= [schedule: '2022-01-04' (Tuesday)] < '2022-01-04 18:00:00')
-
     Args:
-        bounds: A fold tuple ``(start, mid, end)``, or just `start` (followed by `mid` and `end`).
-        mid: Datetime-like. Must be ``None`` when `bounds` is a tuple.
-        end: Datetime-like. Must be ``None`` when `bounds` is a tuple.
+        start_or_bounds: A fold tuple ``(start, mid, end)``, or just `start` (followed by `mid` and `end`).
+        mid: Datetime-like. Must be ``None`` when `start_or_bounds` is a tuple.
+        end: Datetime-like. Must be ``None`` when `start_or_bounds` is a tuple.
         format: A custom format to use. Use :attr:`~.settings.log_split_progress.FOLD_FORMAT` if ``None``, but note that
             only the `start`, `mid` and `end` keys are available to this function.
 
@@ -58,20 +54,41 @@ def to_string(
     Raises:
         TypeError: If an incorrect number of timestamps are given.
 
+    Examples:
+        Sample format output.
+
+        >>> to_string("2021-12-30", "2022-01-04", "2022-01-04 18:00:00")
+        "'2021-12-30' <= [schedule: '2022-01-04' (Tuesday)] < '2022-01-04 18:00:00'"
+
+        The :attr:`default format <.log_split_progress.FOLD_FORMAT>` was used above.
+
+        Using properties. The `delta` is the distance from `mid` formatted by
+        :func:`~rics.strings.format_seconds`. The `date` property returns a :class:`datetime.date` object.
+
+        >>> to_string(
+        ...     ("2021-12-30", "2022-01-04", "2022-01-04 18:00:00"),
+        ...     format="'{start.date}' [-{start.delta}] <= '{mid.iso}' < [+{end.delta}]",
+        ... )
+        "'2021-12-30' [-5d] <= '2022-01-04T00:00:00' < [+18h]"
+
+        The delta is always positive; you must add the sign yourself.
     """
-    if isinstance(bounds, tuple):
+    if isinstance(start_or_bounds, tuple):
         if not (mid is None and end is None):
-            raise TypeError(f"Too many arguments: {(bounds, mid, end)}.")
-        start, mid, end = bounds
+            raise TypeError(f"Too many arguments: {(start_or_bounds, mid, end)}.")
+        start, mid, end = start_or_bounds
     else:
         if mid is None or end is None:
-            raise TypeError(f"Too few arguments: {(bounds, mid, end)}.")
-        start = bounds
+            raise TypeError(f"Too few arguments: {(start_or_bounds, mid, end)}.")
+        start = start_or_bounds
 
-    return (log_split_progress.FOLD_FORMAT if format is None else format).format(
-        start=_PrettyTimestamp(start),
-        mid=_PrettyTimestamp(mid),
-        end=_PrettyTimestamp(end),
+    if format is None:
+        format = log_split_progress.FOLD_FORMAT
+
+    return format.format(
+        start=_PrettyTimestamp(start, mid),
+        mid=_PrettyTimestamp(mid, mid),
+        end=_PrettyTimestamp(end, mid),
     )
 
 
@@ -95,9 +112,11 @@ def stringify(old: Timestamp, *, new: Timestamp | None = None, diff_only: bool =
 
 
 class _PrettyTimestamp:
-    def __init__(self, raw: DatetimeTypes) -> None:
+    def __init__(self, raw: DatetimeTypes, anchor: DatetimeTypes | None = None) -> None:
         self.timestamp = Timestamp(raw)
+        self._anchor = None if anchor is None else Timestamp(anchor)
         self._auto: str | None = None
+        self._delta: str | None = None
 
     def __getattr__(self, name: str) -> Any:
         return getattr(self.timestamp, name)
@@ -122,6 +141,25 @@ class _PrettyTimestamp:
             )
             self._auto = format(timestamp, format_spec)
         return self._auto
+
+    @property
+    def delta(self) -> str:
+        if self._anchor is None:
+            raise ValueError("Delta requires an anchor.")
+
+        if self._delta is None:
+            seconds = (self.timestamp - self._anchor).round("s").total_seconds()
+            self._delta = fmt_sec(abs(seconds))
+
+        return self._delta
+
+    @property
+    def iso(self) -> str:
+        return self.timestamp.isoformat()
+
+    @property
+    def date(self) -> datetime.date:
+        return self.timestamp.date()
 
 
 def format_expanded_limits(