+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `types.SplitProgressExtras` generic on `MetricsType`.
 - Moved `support.create_explorer_link()` to new submodule `time_split.app`.
 - Allow passing `Series/DataFrame` as `plot(available=...)`.
+- The `to_string()` function now supports using _.delta_ and _.iso_ (like _.auto_) in the format.
 
 ## [0.6.0] - 2024-08-31